+++ src/Symfony/Bridge/Twig/Extension/SecurityExtension.php
@@ -14,6 +14,7 @@ namespace Symfony\Bridge\Twig\Extension;
 use Symfony\Component\Security\Acl\Voter\FieldVote;
 use Symfony\Component\Security\Core\Authorization\AccessDecision;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use Symfony\Component\Security\Core\Authorization\GuestAuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Authorization\UserAuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationCredentialsNotFoundException;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -67,7 +68,7 @@ final class SecurityExtension extends AbstractExtension
         return $accessDecision;
     }
 
-    public function isGrantedForUser(UserInterface $user, mixed $attribute, mixed $subject = null, ?string $field = null, ?AccessDecision $accessDecision = null): bool
+    public function isGrantedForUser(?UserInterface $user, mixed $attribute, mixed $subject = null, ?string $field = null, ?AccessDecision $accessDecision = null): bool
     {
         if (null === $this->securityChecker) {
             return false;
@@ -75,6 +76,10 @@ final class SecurityExtension extends AbstractExtension
 
         if (!$this->securityChecker instanceof UserAuthorizationCheckerInterface) {
             throw new \LogicException(\sprintf('You cannot use "%s()" if the authorization checker doesn\'t implement "%s".', __METHOD__, UserAuthorizationCheckerInterface::class));
+        }
+
+        if (null === $user && !$this->securityChecker instanceof GuestAuthorizationCheckerInterface) {
+            throw new \LogicException(\sprintf('You cannot use "%s()" with a null user if the authorization checker doesn\'t implement "%s".%s', __METHOD__, GuestAuthorizationCheckerInterface::class, interface_exists(GuestAuthorizationCheckerInterface::class) ? '' : ' Try upgrading the "symfony/security-core" package to version 8.2 or newer.'));
         }
 
         if (null !== $field) {
@@ -92,7 +97,7 @@ final class SecurityExtension extends AbstractExtension
         }
     }
 
-    public function getAccessDecisionForUser(UserInterface $user, mixed $attribute, mixed $subject = null, ?string $field = null): AccessDecision
+    public function getAccessDecisionForUser(?UserInterface $user, mixed $attribute, mixed $subject = null, ?string $field = null): AccessDecision
     {
         if (!class_exists(AccessDecision::class)) {
             throw new \LogicException(\sprintf('Using the "access_decision_for_user()" function requires symfony/security-core >= 7.3. Try running "%s".', $this->securityChecker ? 'composer update symfony/security-core' : 'composer require symfony/security-core'));